Refactor backend structure and add backup script
Moved main FastAPI logic to use routers for upload and chat endpoints, simplifying main.py. Added backend/backup/main_backup.py as a standalone backup of the previous monolithic backend logic. Removed backend/main_copy.py as it is now redundant.
